langserver: remove fuzzy-matching hover testing
It's just silly for our tests to use `strings.Contains` in order to identify
equality. It's incredibly misleading because I just wrote `wantHover` as `type Foo struct`
and it passes regardless of the fact that the true output is e.g. `type Foo struct{...freakishly long string...}`

Fuzzy-matching in tests is bad!