langserver: make ContainerName output match between `workspace/symbol` and `workspace/reference`
This change causes the `ContainerName` output of both `workspace/symbol` and `workspace/reference`
to fully align with one another. Previously, `workspace/symbol` would output the package name for
top-level functions/vars/etc while `workspace/reference` would not.

Because of the fact that having the package name as the container name when one is not present
helps distinguish between e.g. `NewRouter` defined in `package foo` and `package foo_test` in the
same directory, I've opted to keep that pattern and apply the fix to `workspace/reference`.