langserver: optimize textDocument/symbol method
This optimizes the textDocument/symbol method, in particular for one-off
operations. Prior to this change we would only omit a packages results
after it was imported (via `collectFromPkg`). This was silly, because we
can do the same operation on the package list created by `buildutil.ExpandPatterns`
before even starting a goroutine for `collectFromPkg`. This significantly
reduces the amount of round-tripping needed in terms of VFS operations
to import every Go package in the workspace. Now only the Go package specified
by the file is imported.