langserver: limit workspace/reference to 1/4 CPU by default + add WORKSPACE_REFERENCE_PARALLELISM env var
This limits workspace/reference to using only one fourth of the CPU by default. This is because, currently,
it is quite an intensive operation. Typically the operation is only ran in the background, so it running
slower is OK but not when it is a detriment to every other LSP operation (which is often user-invoked).

Running with `WORKSPACE_REFERENCE_PARALLELISM=8` will retain the old behavior prior to this change.