Add panic handler to worker goroutines
Currently we are getting panics sometimes running workspace/symbols due to a
panic in `go/doc`.

This change prevents panics in our pkg collector goroutines taking the whole
server down. In the future we may want to have a better approach so we don't
accidently ignore panics, but for now this lets us move forward.