references: Cache the reverse import graph (#31)
We compute this every time we answer a references query and for large workspaces
it dominates the time. However, it is static W.R.T. the workspace. In the future
when we start doing cache invalidation, this will need to be updated.

On my machine it takes subsequent references queries from 20s to 3s.