langserver: always emit SymbolDescriptor fields
The spec doesn't make it clear whether or not we should e.g. omit `vendor` if it's value is
`false` (because it has no such concept) but I lean towards the view that we should always
include any fields we _can_ emit, even if their value is the zero value.