langserver: remove invalid TODO about properly handling vendor
Our handling of vendor currently is OK. Today we emit a reference to any Go dependency
(regardless of location, including vendor). This is OK because it means we will treat
`workspace/xreferences` as a reference listing method -- and not do anything fancy like
remove references at the build server (because nobody would ever query for a definition
inside of the repository). Effectively, we handle this at the dependency generation step
instead.