langserver: ctxvfs.NameSpace that is safe for concurrent read/writes
ctxvfs.NameSpace is not safe to use for concurrent read/writes (were a write is
a call to .Bind). As such we create a copy every time we want to use it in a
build.Context + it actually had some race conditions outside of that code. As
part of the perf work I am doing, I want to be able to bind to the NameSpace and
have it instantly appear to all BuildContexts. So this commit introduces
AtomicFS, a wrapper around NameSpace optimized for reads (since Bind calls are
relatively rare). The implementation avoids creating a copy every time we want a
BuildContext (common operation in our build server) + is completely safe for
concurrent access.