langserver: Prevent typecheck cache invalidation race
If we are busy typechecking something and we receive a FS command which mutates
the overlay, we can store in the cache invalid results for future hovers. This
change does not fully prevent that behaviour. However, it does make it much less
likely. The reason it could still happen is the FS tied to the bctx could be
stale. However, the amount of time that lives is low and I will fix it in a
future commit.

This change also simplifies the logic as well as removing the mutex map we had.
This single-flighting pattern is the same as the one used in the `go/loader`
pkg.