langserver: Faster package list (#82)
`buildutil.ExpandPattern` would inspect every package in the `GOPATH` and `GOROOT`.
The slowness of this is not much of a factor on sourcegraph.com, but for local
`GOPATH`s which can be large (such as my own), it is really slow. This takes
symbol search on this project on my machine from half a minute to instant.

This code is essentially the `buildutil.ExpandPattern` modified for our usecase. The idea
used can probably be submitted upstream since there are TODOs hinting at wanting this
functionality.

Previously we also incorrectly relied on `h.init.RootImportPath`, which is
technically not allowed in the langserver. In the future, we should work out
ways to prevent this accidental use of it.

Test Plan: Tested on the extensive integration tests for sourcegraph.com, which
will test both golang/go and other popular pkgs.