symbol: single-flight cache access (#91)
* symbol: Inline pkg cache helpers

This makes the next change clearer.

* symbol: Add single-flighting to cache

When a user types a in a search query, we may start several queries at once. It
is quite common to duplicate work in such cases. This allows us to single-flight
similiar to how we do it for the typecheck cache. Additionally this change also
lets us cache pkgs we failed to parse. Previously we would just repeatedly scan
and fail.

* readability comment