langserver: Process level cache
Since we can potentially have unbounded langservers all running in the same
process, we need to limit the size of the cache at the process level rather than
at the LangHandler level.

This change moves the cache from being managed by LangHandler to being managed
at the process level. However, we hide that from users of the cache by storing
an identifier for a cache in the key.