langserver: Make caches bounded at the process level (#92)
* langserver: Unified cache interface

This encodes our cache + single-flighting pattern into an easier to
use/understand interface. This will also makes it easier to change how we cache.
Soon I plan on making the cache bounded (eviction + process level).

* langserver: Use an LRU cache

We OOM in production due to our unbounded caches. This will make it less likely.
The cache size is arbitrary, future commits will improve this.

* langserver: Process level cache

Since we can potentially have unbounded langservers all running in the same
process, we need to limit the size of the cache at the process level rather than
at the LangHandler level.

This change moves the cache from being managed by LangHandler to being managed
at the process level. However, we hide that from users of the cache by storing
an identifier for a cache in the key.

* langserver: Configurable cache sizes

* cache: Instrument hit/miss in prometheus