textDocument/xdefinition: do not return "nil" AND no error.
The return type signature of `handleDefinition` is `([]lsp.Location, error)`. If `nil, nil`
is returned then ` (*LangHandler).Handle` also returns `result interface{}` as `nil`. When
the result of a jsonrpc2 handler is nil, `jsonrpc2.HandlerWithError` [converts it into a struct](https://sourcegraph.com/github.com/sourcegraph/jsonrpc2/-/blob/handler_with_error.go#L25-28)
which is then encoded as a JSON object instead of an array. When decoding, this results in:

```
json: cannot unmarshal object into Go value of type []lspext.SymbolLocationInformation
```

Which is (obviously) bad because that looks like an error instead of 'no results'.

According to that jsonrpc2 implementation (linked above):

```
// isNilValue tests if an interface is empty, because an empty interface does
// not encode any information, we can't encode it in JSON so that the proxy
// knows it's a response, not a request.
func isNilValue(resp interface{}) bool {
```

This makes it unclear to me whether or not the conversion into a struct is
truly correct or not. I'm open to other ideas on how to resolve this issue.