langserver: Find references defined in tests
`AfterTypeCheck` can be called twice. This is from its docstrings:

  The function may be called twice for the same PackageInfo:
  once for the files of the package and again for the
  in-package test files.

So on the first call we wouldn't find anything defined in-package test files, so
we would set `afterTypeCheckErr`. The second call would then find it, but the
old logic would return the previous error rather than use the found object.