langserver/internal/refs: use per-file best-effort approach
Previously `langserver/internal/refs` would abort on the first error
found in a package. After this change, we make a best-effort approach
to finding references on a per-file level instead of only at the per-package
level at the callsite in `workspace_refs.go`.

This is useful because while not all Go packages can be built, e.g. this one
that I ran into in the wild which imports a broken-path `ntoolkit/component` package:

https://github.com/shadowmint/go-component/blob/master/src/ntoolkit/component/component_fake_component_test.go

We can still find useful information from them, e.g. references to `testing.T`
above. This also occurs in e.g. this kubernetes staging directory:

https://github.com/kubernetes/kubernetes/tree/master/staging

So the use cases for this change are real, and there is no real harm in
making this change.