refs: Stream back partial results (#125)
This is an experiment for streaming back reference results, before the reference
operation is finished. It uses an unspecced notification called
`$/partialResult` which is a list of JSONPatch operations to apply. We stream
back results every second if we have any unsent results.

Currently the order results are sent back in is very different to what the final
result will be. It will also ignore the limit specifier. We can adjust this in
future as we see how the prototype behaves.