hover: Use a nodes Doc commentgroup (#123)
Previously we used `ast.NewCommentMap` which would suck in non-docstring related
comments. An common example of a bad docstring it would produce is `Fprintf`
would include this line at the start of the docstring
https://sourcegraph.com/github.com/golang/go@6bdb0c11c73ecf2337918d784c54f9dda2207ca7/-/blob/src/fmt/print.go#L175:1-176:1

The test case was updated to reflect this behaviour. Before this change the test
case fails.