langserver: Publish empty diagnostics
We publish empty diagnostics if we had a typecheck cache miss (ie we actually
did typechecking). This change works since at this point diags is nil
https://sourcegraph.com/github.com/sourcegraph/go-langserver@39e41a9f7071d047fc97dfe52d51dca11d6ac04b/-/blob/langserver/loader.go#L197:1-198:1
but is only updated to the diags returned by `typecheck` if we have a cache
miss. `typecheck` always returns a non-nil (but possibly empty) diags after this
change.