tools: Only search for references under rootPath
This should be a decent perf win on large GOPATHs. Currently we would search the
whole GOPATH to build up the import graph, but only do queries inside of the
workspace root. That means all those extra edges/nodes are never touched. From
testing, building the importgraph is usually the slowest part of reference
requests.

This also has the added benefit of adding support for xtests.