references: Perf improvements to importgraph building (#132)
* langserver: Move listPkgUnderDir to internal tools pkg

We will start using this pkg more for storing functionality that is closely
related to code from golang.org/x/tools

* tools: fork importgraph.Build

This is a direct copy, so the changes I make are clearer.

* tools: Only return reverse import graph

We don't use the forward graph, nor do we handle the errors.

* tools: Use custom FindPackageFunc for importgraph

This allows us to use our custom find package func if it is available. On
sourcegraph.com it is significantly faster and tackles many of the problems
present in the perf TODO also removed in this commit.

* tools: Only search for references under rootPath

This should be a decent perf win on large GOPATHs. Currently we would search the
whole GOPATH to build up the import graph, but only do queries inside of the
workspace root. That means all those extra edges/nodes are never touched. From
testing, building the importgraph is usually the slowest part of reference
requests.

This also has the added benefit of adding support for xtests.