references: Major refactoring
The main goal I want to achieve is to start doing a typecheck with a potentially
out of date ImportGraph, and then follow up with typechecking an up to date
ImportGraph. To avoid doing extra work + being able to do it twice cleanly
required a bunch of refactoring. The next change should be clearer now.

My only concern in this change is the switch from collecting refs in
`[]*ast.Ident` batches, to sending each individual `*ast.Ident` down a channel.
This may have a negative impact on perf, but in practice I believe this will be
minor. In future we will only stream references as well, so this will be fine
when we don't have to build up the final result.