references: Make it possible to stream import graphs
We are going to adjust the implementation to start typechecking packages sooner
in the hope we can start returning results sooner to the user. For now this is a
no-op change w.r.t. perf and output.

It does fix a bug which does not happen yet. We used the request context for the
import graph previously. If that respected timeouts, we would cache an
incomplete import graph. However, that was not the case currently.