Some refactoring to overlay fs (#143)
* fs: Remove OverlayMountPath

We only ever wrote to it, never read it. It was also always RootPath. We have
one read in our build server, but that can be replaced with RootPath.

* fs: Refactor overlay into own struct

This makes the responsibility of overlay much more clear. We also get a much
easier to test interface as well as less lock contention. Previously we would
lock h.Mu and h.overlayMu to modify the overlay. Now we only need to lock the
latter.