fs: Only invalidate cache if files change (#145)
* fs: Only invalidate cache if we modify the overlay

* fs: Check if we change a file before invalidating cache

We want this behaviour so that we can start forwarding didOpen requests in
production and correctly handle them with respect to caches.

* fs: Typecheck modified/opened files

If we receive a FS request, we know the user is interacting with the file. So we
can use this information to start typechecking the file ahead of time.