fs: Use version field to resolve out of order FS updates (#160)
If we have quickly firing didChange notifs, the order they are processed in
can start being racey depending on the order the goroutine for each notif are
run. As such an older didChange for a document can be processed after a
newer didChange. Since we only advertise full syncing, we can rely on only using
didChange requests if they are replacing a contents with an older version
number.