do not emit a containerName in the SymbolInformation for package-scope defs
vscode's internal `skipLocalSymbols` option treats SymbolInformations with containerName set as locals, which are skipped. This means that when that option is used, there are no Go symbol results.

This fixes our interpretation of containerName to match vscode's assumptions.

There is minimal end-user impact. If there is no containerName, vscode shows the file path, which is actually more useful than the containerName we previously used (the package name).