allow callers to indicate they do not need deps to be fetched when they call FindPackage
Operations that only require parsing of a single file (or a single package's files) do not need deps to be fetched. Skipping dep fetching makes them much faster.

Use case: The gitlens vscode extension calls textDocument/documentSymbol on each document when it is opened, to know where to position its code lenses containing git information. This should be a super quick parsing-only task, but it takes a long time when using the Go lang server because it must fetch deps needlessly.