textDocument/symbols: Directly parse document
Previously we would just filter the output of workspace/symbol to support text
document symbol. However, we have extra logic related to avoiding conflicting
packages. That logic is convoluted and needs to be simplified, but care needs to
be taken for that. This PR factors out some of the core symbol collection logic
so that it is easy to get LSP symbols for an AST.