Use 'godef'-based implementation for utilizing $GOPATH/pkg .a files (improves performance)
This change adds a `-usebinarypkgcache=true` flag which uses `$GOPATH/pkg` binary `.a`
files in order to improve the performance of `textDocument/definition` requests. In short,
we no longer require typechecking to perform these requests which saves us several hundreds
of MB of memory _and_, more importantly, a lot of CPU resources.

Because this uses .a files from disk, this option will be disabled on sourcegraph.com where
only the VFS is used. In the future, however, we can and should consider consolidating these
two code paths where possible.

Our existing test suite has proven very useful during development in ensuring this
implementation is at least on par with our current one, so there is no concern of this being
sub-par or broken in subtle ways as far as I see.