langserver: fix definition & hover on unsaved files
This change makes it such that hovering or jumping to definition while
inside a file that has been modified always works. There is more to be
done in the future, in this regard, for e.g. hover on a modified
dependency [see here](https://github.com/sourcegraph/go-langserver/pull/196#discussion_r123695661)
but this change fixes the most common case (hovering/jumping to def on
code that is actively being written / modified).