Vendor dependencies
We are mainly consumed as a command, so we should prevent future breakages and
vendor our changes in. The main consumer of us as a library is Sourcegraph, but
dep should flatten the deps in that case. We also declare no constraints in
Gopkg.toml to prevent it influencing anyone who uses go-langserver as a
library.