fix uri handling for windows
the main problem was that some parts of the code converted the path to
lower case and others just passed on the path as it was. e.g
didOpen/didChange messages just used the path whereas hover used the
lower case path to read the file, so it never read the changed file from
the overlay cache!