Reorder log/fatal in langserver_test.
Under circumstances I have not adequately debugged, the
	GOPATH=[tmpdir] go install -v all
used to set up test cases can fail. In my particular case, this appears
to be because something has convinced it that it absolutely has to
rewrite cmd/pprof in $GOROOT.

Reordering these (and adding the GOPATH=...) means that, if the test
fails for this reason, we get a diagnostic more specific than "exit code 1".

Signed-off-by: Seebs <seebs@seebs.net>