Improve type definition lookup
Handle type lookup more cleanly; in particular, since we can
only work on Named types anyway, instead of doing fancy scope
and package lookups, just rely on the named type having a
usable name.

Also, handle pointer-to-named-type (or pointer-to-pointer-to,
etcetera) cleanly.

This introduces a diagnostic from one of the test cases
which I haven't debugged yet, but it passes the test cases
even if it produces extra logging.

Signed-off-by: Seebs <seebs@seebs.net>