Improve handling of slices and pointers
Dereference slices and pointers recursively. Also make parallel
change in internal/refs. I don't think it makes sense to try to
implement the same thing for maps, so the diagnostic from refs
for the map test case is probably correct, and I'm not sure how
or whether to fix it.

Signed-off-by: Seebs <seebs@seebs.net>