Initial implementation for textdocument/typeDefinition (#264)
* Draft implementation for textdocument/typeDefinition
This adds a test case for textdocument/typeDefinition, and also
an initial implementation. The implementation does not attempt
to handle the Godef case (yet?), because that case appears to
be an optional optimization.

Passes 'go test', anyway.

* Improve type definition lookup

Handle type lookup more cleanly; in particular, since we can
only work on Named types anyway, instead of doing fancy scope
and package lookups, just rely on the named type having a
usable name.

Also, handle pointer-to-named-type (or pointer-to-pointer-to,
etcetera) cleanly.

This introduces a diagnostic from one of the test cases
which I haven't debugged yet, but it passes the test cases
even if it produces extra logging.

Signed-off-by: Seebs <seebs@seebs.net>

* Improve handling of slices and pointers

Dereference slices and pointers recursively. Also make parallel
change in internal/refs. I don't think it makes sense to try to
implement the same thing for maps, so the diagnostic from refs
for the map test case is probably correct, and I'm not sure how
or whether to fix it.

Signed-off-by: Seebs <seebs@seebs.net>

* cleanup type-dereferencing logic

Further cleanup of type-dereferencing logic; now handles all
of array, slice, pointer, and channel in comparable ways, and
does it with iteration rather than recursion, saving what
could easily be several nanoseconds for deeply nested types.

Signed-of-by: Seebs <seebs@seebs.net>

* Demethodize a non-method

h.typeLookup did not actually refer to h in any way, it is not
a method.

Signed-off-by: Seebs <seebs@seebs.net>

* Slight cleanup, fix fencepost

Had spurious +1 in length calculation, and tests which tested for the
corresponding wrong value. Nice! Fixed that.

Drop a no-longer-relevant comment, and a variable declaration which
did not add clarity.

Signed-off-by: Seebs <seebs@seebs.net>