pkg/lsp: Move to external repository (#325)
We use these types in multiple repositories, which introduces versioning issues.
Currently we version this repository like you would a command, increasing the
major versions without respect to semver constraints on the libraries it
exports. With go modules this is frustrating since we need to update the import
paths every time we do major version bump. So we have moved the lsp structs into
another repository.

To keep backwards compatability we have added type aliases to forward the old
types to the new types.