add support for zipURLTemplate in initializationOptions
This is used when fetching dependencies as a performance optimization
over fetching from GitHub's codeload endpoint or a full git clone.

The client sends the template to go-langserver, and go-langserver will
try fetching the repo contents from the URL built by instantiating the
template with the repo name and revision of the dependency.

The template must have 2 `%s`s: one for the repo name and one for the
revision.