dockerfile: use builder pattern to reduce image size
Docker image size goes from 625mb to 67mb since we exclude the build
environment and source. We also include tini and bind-tools which is standard
practice in our images now.

There is a risk we now act slightly worse due to not including the go build
toolchain. I will validate this and rollback this change if that is indeed the
case.